code refactoring and cleanup for export
