+++ src/HackStreetCLIExtensions/CommandHandlers/ExportCommandHandler.cs
@@ -1,44 +1,21 @@
 ﻿using HackStreetCLIExtensions.Extensions;
 using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Newtonsoft.Json;
 using Sitecore.CH.Cli.Core.Abstractions.Commands;
 using Sitecore.CH.Cli.Core.Abstractions.Rendering;
-using SitecoreCHCLIExtensions.Models;
-using Stylelabs.M.Base.Querying;
+using HackStreetCLIExtensions.Models;
 using Stylelabs.M.Base.Querying.Filters;
 using Stylelabs.M.Framework.Essentials.LoadConfigurations;
 using Stylelabs.M.Sdk.Contracts.Base;
 using Stylelabs.M.Sdk.Contracts.Querying;
 using Stylelabs.M.Sdk.WebClient;
 using System.CommandLine.Invocation;
-using System.Dynamic;
 using System.Globalization;
-using static System.Runtime.InteropServices.JavaScript.JSType;
+using Stylelabs.M.Base.Querying;
 
-namespace SitecoreCHCLIExtensions.Export
+namespace HackStreetCLIExtensions.CommandHandlers
 {
-    public class ExportCommand : BaseCommand<ExportCommandHandler>
-    {
-        public ExportCommand() : base("export", "Export Assets")
-        {
-            AddOption<string>(ExportMessages.ExportCommandQuery, false, new string[2]
-            {
-                "--query",
-                "-q"
-            });
-            AddOption<string>(ExportMessages.ExportCommandLocation, false, new string[2]
-            {
-                "--location",
-                "-l"
-            });
-            AddOption<string>(ExportMessages.ExportCommandFields, false, new string[2]
-            {
-                "--fields",
-                "-f"
-            });
-        }
-    }
     public class ExportCommandHandler : BaseCommandHandler
     {
         public ExportCommandHandler(Lazy<IWebMClient> client, IOutputRenderer renderer, IOptions<ExportParameters> parameters) : base(client, renderer)
@@ -61,7 +38,7 @@ namespace SitecoreCHCLIExtensions.Export
             if (assetDefinition != null)
             {
                 Renderer.WriteLine("Asset Id 1");
-                var queryFilters = Parameters.Query.Split(",");
+                var queryFilters = Parameters.Query.Split("&");
                 var entityPropertyDefinitions = assetDefinition.GetPropertyDefinitions();
                 var entityRelationDefinitions = assetDefinition.GetRelationDefinitions();
                 if (queryFilters.Length > 0)
@@ -110,7 +87,7 @@ namespace SitecoreCHCLIExtensions.Export
                     };
                     IEntityIterator iterator = client.Querying.CreateEntityIterator(assetQuery, EntityLoadConfiguration.Full);
                     List<JObject> assetEntities = new List<JObject>();
-                    var fieldsArray = fields.Split(',');
+                    var fieldsArray = fields.Split('|');
                     HashSet<string> excelKeys = new HashSet<string>();
                     excelKeys.Add("Identifier");
                     excelKeys.Add("File");
@@ -123,95 +100,101 @@ namespace SitecoreCHCLIExtensions.Export
                         {
                             Renderer.WriteLine("Asset Id :");
                             Renderer.WriteLine(entity.Id.ToString());
-                            JObject assetEntityObj = new JObject();
-                            var entityIdentifier = entity.Identifier;
                             var publicLink = ContentHubHelperExtension.FetchAssetPublicLink(client, entity, Renderer);
-                            Renderer.WriteLine("Public Link :");
-                            Renderer.WriteLine(publicLink);
-                            assetEntityObj["Identifier"] = entityIdentifier;
-                            assetEntityObj["File"] = publicLink;
-                            foreach (var fieldProp in fieldsArray)
-                            {
-                                var isPropertyField = entityPropertyDefinitions.Where(x => x.Name == fieldProp)?.FirstOrDefault();
-                                var isRelationalField = entityRelationDefinitions.Where(x => x.Name == fieldProp)?.FirstOrDefault();
-
-                                if (isPropertyField != null)
+                            if (!string.IsNullOrEmpty(publicLink)){
+                                Renderer.WriteLine("Public Link :");
+                                Renderer.WriteLine(publicLink);
+                                var entityIdentifier = entity.Identifier;
+                                JObject assetEntityObj = new JObject();
+                                assetEntityObj["Identifier"] = entityIdentifier;
+                                assetEntityObj["File"] = publicLink;
+                                foreach (var fieldProp in fieldsArray)
                                 {
-                                    if (isPropertyField.IsMultiLanguage)
+                                    var isPropertyField = entityPropertyDefinitions.Where(x => x.Name == fieldProp)?.FirstOrDefault();
+                                    var isRelationalField = entityRelationDefinitions.Where(x => x.Name == fieldProp)?.FirstOrDefault();
+
+                                    if (isPropertyField != null)
                                     {
-                                        Renderer.WriteLine("is Multi");
-                                        var excelfieldName = $"{fieldProp}#en-US";
-                                        if (!excelKeys.Contains(excelfieldName))
+                                        if (isPropertyField.IsMultiLanguage)
                                         {
-                                            excelKeys.Add(excelfieldName);
-                                        }
-                                       
-                                        Renderer.WriteLine(fieldProp);
-                                        string propertyValue = entity.GetPropertyValue<string>(fieldProp, CultureInfo.GetCultureInfo("en-US"));
-                                        if (!string.IsNullOrEmpty(propertyValue))
-                                        {
-                                            Renderer.WriteLine(propertyValue);
-                                            assetEntityObj[excelfieldName] = propertyValue;
-                                            Renderer.WriteLine("is MultiCompleted");
+                                            Renderer.WriteLine("is Multi");
+                                            var excelfieldName = $"{fieldProp}#en-US";
+                                            if (!excelKeys.Contains(excelfieldName))
+                                            {
+                                                excelKeys.Add(excelfieldName);
+                                            }
+
+                                            Renderer.WriteLine(fieldProp);
+                                            string propertyValue = entity.GetPropertyValue<string>(fieldProp, CultureInfo.GetCultureInfo("en-US"));
+                                            if (!string.IsNullOrEmpty(propertyValue))
+                                            {
+                                                Renderer.WriteLine(propertyValue);
+                                                assetEntityObj[excelfieldName] = propertyValue;
+                                                Renderer.WriteLine("is MultiCompleted");
+                                            }
+                                            else
+                                            {
+                                                assetEntityObj[excelfieldName] = string.Empty;
+                                            }
                                         }
                                         else
                                         {
-                                            assetEntityObj[excelfieldName] = string.Empty;
+                                            Renderer.WriteLine("Non multi");
+                                            var excelfieldName = $"{fieldProp}";
+                                            if (!excelKeys.Contains(excelfieldName))
+                                            {
+                                                excelKeys.Add(excelfieldName);
+                                            }
+                                            string propertyValue = entity.GetPropertyValue<string>(fieldProp);
+                                            if (!string.IsNullOrEmpty(propertyValue))
+                                            {
+                                                Renderer.WriteLine(propertyValue);
+                                                assetEntityObj[excelfieldName] = propertyValue;
+                                                Renderer.WriteLine("Non multi MultiCompleted");
+                                            }
+                                            else
+                                            {
+                                                assetEntityObj[excelfieldName] = string.Empty;
+                                            }
                                         }
                                     }
-                                    else
+                                    else if (isRelationalField != null)
                                     {
-                                        Renderer.WriteLine("Non multi");
+                                        Renderer.WriteLine("Relational Id 1");
                                         var excelfieldName = $"{fieldProp}";
                                         if (!excelKeys.Contains(excelfieldName))
                                         {
                                             excelKeys.Add(excelfieldName);
                                         }
-                                        string propertyValue = entity.GetPropertyValue<string>(fieldProp);
-                                        if (!string.IsNullOrEmpty(propertyValue))
+                                        IRelation entityRelation = entity.GetRelation(fieldProp);
+                                        List<string> relationalValues = new List<string>();
+                                        if (entityRelation != null)
                                         {
-                                            Renderer.WriteLine(propertyValue);
-                                            assetEntityObj[excelfieldName] = propertyValue;
-                                            Renderer.WriteLine("Non multi MultiCompleted");
-                                        }
-                                        else
-                                        {
-                                            assetEntityObj[excelfieldName] = string.Empty;
-                                        }
-                                    }
-                                }
-                                else if (isRelationalField != null)
-                                {
-                                    Renderer.WriteLine("Relational Id 1");
-                                    var excelfieldName = $"{fieldProp}";
-                                    if (!excelKeys.Contains(excelfieldName))
-                                    {
-                                        excelKeys.Add(excelfieldName);
-                                    }
-                                    IRelation entityRelation = entity.GetRelation(fieldProp);
-                                    List<string> relationalValues = new List<string>();
-                                    if (entityRelation != null)
-                                    {
-                                        var relationalIds = entityRelation.GetIds();
-                                        if (relationalIds != null && relationalIds.Any())
-                                        {
-                                            foreach (var relationalId in relationalIds)
+                                            var relationalIds = entityRelation.GetIds();
+                                            if (relationalIds != null && relationalIds.Any())
                                             {
-                                                var relationalEntity = client.Entities.GetAsync(relationalId).ConfigureAwait(false).GetAwaiter().GetResult();
-                                                if (relationalEntity != null)
+                                                foreach (var relationalId in relationalIds)
                                                 {
-                                                    relationalValues.Add(relationalEntity.Identifier);
+                                                    var relationalEntity = client.Entities.GetAsync(relationalId).ConfigureAwait(false).GetAwaiter().GetResult();
+                                                    if (relationalEntity != null)
+                                                    {
+                                                        relationalValues.Add(relationalEntity.Identifier);
+                                                    }
                                                 }
+                                                Renderer.WriteLine(string.Join("|", relationalValues));
+                                                assetEntityObj[fieldProp] = string.Join("|", relationalValues);
                                             }
-                                            Renderer.WriteLine(string.Join("|", relationalValues));
-                                            assetEntityObj[fieldProp] = string.Join("|", relationalValues);
                                         }
                                     }
                                 }
+                                Renderer.WriteLine("JSON");
+                                Renderer.RenderJson(assetEntityObj);
+                                assetEntities.Add(assetEntityObj);
                             }
-                            Renderer.WriteLine("JSON");
-                            Renderer.RenderJson(assetEntityObj);
-                            assetEntities.Add(assetEntityObj);
+                            else
+                            {
+                                Renderer.WriteLine("Public Link does not exist");
+                            }
                         }
                         Renderer.RenderJson(assetEntities);
                         Renderer.RenderJson(excelKeys);
