--- src/HackStreetCLIExtensions/Import/ImportMessages.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HackStreetCLIExtensions.Import
-{
-    internal class ImportMessages
-    {
-        public const string ImportCommandLocation = "Location to export file";
-    }
-}
