+++ src/HackStreetCLIExtensions/Messages/ImportMessages.cs
@@ -1,0 +1,7 @@
+﻿namespace HackStreetCLIExtensions.Messages
+{
+    internal class ImportMessages
+    {
+        public const string ImportCommandLocation = "Location to export file";
+    }
+}
